cleanするときは確認ダイアログ出すようにした。renumberをmakeにくみこんだ

+++ Makefile
@@ -2,9 +2,19 @@
 IMAGE_DIR := ./screencapture
 FFMPEG := ffmpeg
 DATE := $(shell date +'%Y%m%d')
-# TARGETS := $(wildcard ./screencapture/*.png)
 
 all:
+	/usr/bin/php renumber.php $(IMAGE_DIR) $(DATE)
 	$(FFMPEG) -r 10 -i "$(IMAGE_DIR)/$(DATE)_%05d.png" -pix_fmt yuv420p -vcodec h264 -s 1120x700 -y "worklog_$(DATE).mp4"
 clean:
+	@while [ -z "$$CONTINUE" ]; do \
+		read -r -p "Type anything but Y or y to exit. [y/N] " CONTINUE; \
+	done ; \
+	if [ ! $$CONTINUE == "y" ]; then \
+	if [ ! $$CONTINUE == "Y" ]; then \
+		echo "Exiting." ; exit 1 ; \
+	fi \
+	fi
 	rm $(IMAGE_DIR)/$(DATE)_*.png
+rename:
+	rename.sh $(DATE) $(IMAGE_DIR)
